namespacing refactor
