+++ src/Entity/CallerInterface.php
@@ -3,16 +3,16 @@
  * Licensed under The GPL-3.0 License
  * For full copyright and license information, please see the LICENSE.txt
  * Redistributions of files must retain the above copyright notice.
- * 
+ *
  * @author   Christopher Castro <chris@quickapps.es>
  * @license  http://opensource.org/licenses/gpl-3.0.html GPL-3.0 License
  */
-namespace Bingo;
+namespace Bingo\Entity;
 
 /**
- * Represents a single player within a bingo game session.
+ * Represents a bingo caller.
  */
-interface PlayerInterface
+interface CallerInterface
 {
 
 }
